+++ src/GHS3DPlugin/GHS3DPlugin_Hypothesis.hxx
@@ -134,14 +134,14 @@ public:
   /*!
    * Maximal size of memory to be used by the algorithm (in Megabytes)
    */
-  void SetMaximumMemory(long MB);
-  long GetMaximumMemory() const;
+  void SetMaximumMemory(float MB);
+  float GetMaximumMemory() const;
   /*!
    * Initial size of memory to be used by the algorithm (in Megabytes) in
    * automatic memory adjustment mode. Default is zero
    */
-  void SetInitialMemory(long MB);
-  long GetInitialMemory() const;
+  void SetInitialMemory(float MB);
+  float GetInitialMemory() const;
   /*!
    * Optimization level: 0-none, 1-light, 2-medium, 3-standard+, 4-strong. Default is medium
    */
@@ -290,8 +290,8 @@ public:
   
   static bool   DefaultMeshHoles();
   static bool   DefaultToMakeGroupsOfDomains();
-  static long   DefaultMaximumMemory();
-  static long   DefaultInitialMemory();
+  static float  DefaultMaximumMemory();
+  static float  DefaultInitialMemory();
   static short  DefaultOptimizationLevel();
   static std::string DefaultWorkingDirectory();
   static bool   DefaultKeepFiles();
@@ -338,8 +338,8 @@ protected:
 
   bool        myToMeshHoles;
   bool        myToMakeGroupsOfDomains;
-  long        myMaximumMemory;
-  long        myInitialMemory;
+  float       myMaximumMemory;
+  float       myInitialMemory;
   short       myOptimizationLevel;
   bool        myKeepFiles;
   std::string myWorkingDirectory;
